Updated shrapnel
Take note, shrapnel seem like it's working when it collides with the 2 top right temp enemies, that's because some shrapnel immediately collides with it again and thus it dissappear. So, it's working as intended.

The shrapnel does not work with the other enemy array yet, only works with the temp enemy, because the enemy array is incomplete